consolidated routes folder and controllers folder
