+++ .env
@@ -1,3 +1,3 @@
 DB_NAME='blog_db'
 DB_USER='root'
-DB_PASSWORD='zelda12345'
+DB_PASSWORD=''
