+++ config/config.json
@@ -1,21 +1,21 @@
 {
   "development": {
     "username": "root",
-    "password": "zelda12345",
+    "password": "",
     "database": "blog_db",
     "host": "127.0.0.1",
     "dialect": "mysql"
   },
   "test": {
     "username": "root",
-    "password": "zelda12345",
+    "password": "",
     "database": "database_test",
     "host": "127.0.0.1",
     "dialect": "mysql"
   },
   "production": {
     "username": "root",
-    "password": "zelda12345",
+    "password": "",
     "database": "database_production",
     "host": "127.0.0.1",
     "dialect": "mysql"
