--- public/explore.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Blog</title>
-    <link href="reset.css" rel="stylesheet" />
-    <link href="output.css" rel="stylesheet" />
-    <script src="https://cdn.tailwindcss.com"></script>
-  </head>
-  <body>
-    <navbar
-      class="flex justify-between items-center p-7 border-solid border-4 border-indigo-500/100 border-x-indigo-500 bg-black"
-    >
-      <div class="img">
-        <img src="logo.png" alt="logo" width="150" height="150" />
-      </div>
-      <div id="btn" class="text-white">
-        <a href="login.html" class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">Login</a>
-        <a href="register.html" class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">Register</a>
-      </div>
-    </navbar>
-    <section class="h-screen flex flex-col justify-start max-sm:flex max-sm:flex-col max-md:flex max-md:flex-col">
-      <aside
-      class="p-7 border-solid border-4 border-indigo-500/100 border-x-indigo-500 bg-black text-white flex-none "
-      >
-        <a href="index.html" class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">Home</a>
-        <a href="explore.html" class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">Explore Blogs</a>
-      </aside>
-      <article
-      class="p-7 border-solid border-4 border-indigo-500/100 border-x-indigo-500 bg-red-600 flex-auto"
-      >
-        <p>links to various blogs (dynamic Handlebars content)</p>
-      </article>
-    </section>
-    <footer
-      class="p-7 border-solid border-4 border-indigo-500/100 border-x-indigo-500 bg-red-600"
-    >
-      <p>Copyright</p>
-    </footer>
-  </body>
-</html>
