--- public/register.handlebars
@@ -1,47 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href="/styles/styles.css" rel="stylesheet">
-    <title>Register</title>
-</head>
-
-<body>
-
-<section class="flex justify-start">
-    <aside class="p-7 border-solid border-4 border-indigo-500/100 border-x-indigo-500 bg-black text-white flex-none ">
-        <a href="/" class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">Home</a>
-        <a href="/explore" class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">Explore Blogs</a>
-    </aside>
-    <main class="flex-grow">
-        <form action="/register" method="post" class="p-8 space-y-4">
-            <div>
-                <label for="username" class="block mb-2 text-sm">Username</label>
-                <input type="text" id="username" name="username" required>
-            </div>
-            <div>
-                <label for="firstName" class="block mb-2 text-sm">First Name</label>
-                <input type="text" id="firstName" name="firstName" required>
-            </div>
-            <div>
-                <label for="lastName" class="block mb-2 text-sm">Last Name</label>
-                <input type="text" id="lastName" name="lastName" required>
-            </div>
-            <div>
-                <label for="email" class="block mb-2 text-sm">Email</label>
-                <input type="email" id="email" name="email" required>
-            </div>
-            <div>
-                <label for="password" class="block mb-2 text-sm">Password</label>
-                <input type="password" id="password" name="password" required>
-            </div>
-            <button type="submit" class="block w-full mt-4 bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">Register</button>
-        </form>
-    </main>
-</section>
-
-</body>
-
-</html>
