--- routes/api/index.js
@@ -1,7 +1,0 @@
-const router = require('express').Router();
-
-const userRoutes = require('./UserRoutes');
-
-router.use('/users', userRoutes);
-
-module.exports = router;
