+++ server.js
@@ -6,7 +6,7 @@ const cors = require('cors');
 const exphbs = require('express-handlebars');
 const SequelizeStore = require('connect-session-sequelize')(session.Store);
 
-const routes = require('./routes');
+const routes = require('./controllers');
 const sequelize = require('./config/connection');
 
 const app = express();
