--- test-exphbs.js
@@ -1,3 +1,0 @@
-//test-exphbs.js
-const exphbs = require('express-handlebars');
-console.log(typeof exphbs);
